requirements.txt
annotated-types==0.7.0
anyio==4.10.0
fastapi==0.116.1
idna==3.10
pydantic==2.11.7
pydantic_core==2.33.2
sniffio==1.3.1
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.23.2
python-multipart==0.0.7
Pillow==11.3.0
